Refactor to multi bundle setup
