--- src/loader-surveys-preview.ts
@@ -1,1 +1,0 @@
-export { renderSurveysPreview, renderFeedbackWidgetPreview } from './extensions/surveys'
