+++ src/loaders/loader-exception-autocapture.ts
@@ -1,6 +1,6 @@
-import { extendPostHog } from './extensions/exception-autocapture'
+import { extendPostHog } from '../extensions/exception-autocapture'
 
-import { window } from './utils/globals'
+import { window } from '../utils/globals'
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
