+++ src/loaders/loader-module.ts
@@ -1,0 +1,7 @@
+import { PostHogCore, POSTHOG_INSTANCES } from '../posthog-core'
+import { init_as_module } from '../loaders/loader-helpers'
+
+export * from '../types'
+export * from '../posthog-surveys-types'
+export const posthog = init_as_module(PostHogCore, POSTHOG_INSTANCES)
+export default posthog
