+++ src/loaders/loader-surveys-preview.ts
@@ -1,0 +1,1 @@
+export { renderSurveysPreview, renderFeedbackWidgetPreview } from '../extensions/surveys'
