+++ src/loaders/loader-surveys.ts
@@ -1,6 +1,6 @@
-import { generateSurveys } from './extensions/surveys'
+import { generateSurveys } from '../extensions/surveys'
 
-import { window } from './utils/globals'
+import { window } from '../utils/globals'
 
 if (window) {
     ;(window as any).extendPostHogWithSurveys = generateSurveys
